requirements.txt
schedule
websocket-client
